portico: Move portico view code to its own file.
This improves the readability of the codebase.
