+++ zerver/views/users.py
@@ -1,10 +1,9 @@
 from typing import Union, Optional, Dict, Any, List
 
-import ujson
 from django.http import HttpRequest, HttpResponse
 
 from django.utils.translation import ugettext as _
-from django.shortcuts import redirect, render
+from django.shortcuts import redirect
 from django.conf import settings
 
 from zerver.decorator import require_realm_admin, require_member_or_admin
@@ -23,7 +22,6 @@ from zerver.lib.exceptions import CannotDeactivateLastUserError
 from zerver.lib.integrations import EMBEDDED_BOTS
 from zerver.lib.request import has_request_variables, REQ
 from zerver.lib.response import json_error, json_success
-from zerver.lib.storage import static_path
 from zerver.lib.streams import access_stream_by_name
 from zerver.lib.upload import upload_avatar_image
 from zerver.lib.users import get_api_key
@@ -492,18 +490,3 @@ def get_profile_backend(request: HttpRequest, user_profile: UserProfile) -> Http
         result['max_message_id'] = messages[0].id
 
     return json_success(result)
-
-def team_view(request: HttpRequest) -> HttpResponse:
-    with open(static_path('generated/github-contributors.json')) as f:
-        data = ujson.load(f)
-
-    return render(
-        request,
-        'zerver/team.html',
-        context={
-            'page_params': {
-                'contrib': data['contrib'],
-            },
-            'date': data['date'],
-        },
-    )
